// backend/actions/conflitFrontalier/types.ts
export enum ConflitFrontalierState {
    UNKNOWN = 'UNKNOWN',

    AVENTURE_MENU = 'AVENTURE_MENU',
    AVENTURE_MENU_DEFI_TAB = 'AVENTURE_MENU_DEFI_TAB',

    CONFLIT_POPUP_MATCHMAKING = 'CONFLIT_POPUP_MATCHMAKING',
    CONFLIT_POPUP_TO_ACCEPT = 'CONFLIT_POPUP_TO_ACCEPT',

    CONFLIT_TRANSMISSION_SCREEN = 'CONFLIT_TRANSMISSION_SCREEN',
    CONFLIT_LOADING_SCREEN = 'CONFLIT_LOADING_SCREEN',

    CONFLIT_INSTANCE_IN_PROGRESS = 'CONFLIT_INSTANCE_IN_PROGRESS',
    CONFLIT_INSTANCE_FINISHED = 'CONFLIT_INSTANCE_FINISHED',
    CONFLIT_INSTANCE_EXIT = 'CONFLIT_INSTANCE_EXIT',

    CONFLIT_INSTANCE_EXITED = 'CONFLIT_INSTANCE_EXITED',
}